chainlit_app/requirements.txt
# llamaindex
llama-index==0.12.10
llama-index-core==0.12.10.post1

# Groq AI
llama-index-llms-groq==0.3.1

# Cohere AI Embeddings
llama-index-embeddings-cohere==0.4.0

# Redis Chat store
llama-index-storage-chat-store-redis==0.4.0

# Qdrant
llama-index-vector-stores-qdrant==0.4.3
fastembed==0.5.0

# Chainlit
chainlit==2.1.0

# Phoenix trace
arize-phoenix-otel==0.6.1
openinference-instrumentation-llama-index==3.1.3

# Postgres database
asyncpg==0.30.0
boto3==1.35.95